separate out the class report from a particular database subfolder
