+++ educator_dashboard/class_report.py
@@ -1,18 +1,18 @@
 import pandas as pd
-from .nested_dataframe import flatten
-from .State import State
-from .Query import QueryCosmicDSApi
-from ..database_new.NewState import State as NewState
+from .database.nested_dataframe import flatten
+from .database.State import State
+from .database.Query import QueryCosmicDSApi
+from .database_new.NewState import State as NewState
 import time
 HUBBLE_ROUTE_PATH = "hubbles_law"
 
 import astropy.units as u
 from math import nan
 
-from .utils import l2d, convert_column_of_dates_to_datetime, get_or_none
+from .database.utils import l2d, convert_column_of_dates_to_datetime, get_or_none
 
 from typing import List, Dict, cast, Optional, Any, Union, TypeVar, overload, TYPE_CHECKING, TypedDict
-from .types import (
+from .database.types import (
     StateInterface, 
     OldRosterEntry, 
     NewRosterEntry as ImportedNewRosterEntry,
@@ -22,7 +22,7 @@ from .types import (
     ProcessedState as ImportedProcessedState,
     OldStudentStoryState
 )
-from ..database_new.types import (
+from .database_new.types import (
     NewStudentStoryState,
     StudentEntry as NewRosterEntry,
     NewRoster,
@@ -31,7 +31,9 @@ from ..database_new.types import (
     AppState
 )
 
-from ..logging import logger
+from .cr_types import HubbleData, ProgressSummary, QuestionInfo, MeasurementStatus
+
+from .logging import logger
 
 class Student():
     
@@ -305,7 +307,7 @@ class Roster():
 
         return self.class_summary
     
-    def class_measurement_status(self, refresh = False):
+    def class_measurement_status(self, refresh = False) -> MeasurementStatus:
         g = cast(pd.DataFrame, self.get_class_data(df=True, refresh = refresh)).groupby('student_id')
         nodist = g.apply(lambda x:  5-sum(x['est_dist_value'] == 0)) # number of distances
         novel = g.apply(lambda x:  5-sum(x['velocity_value'] == 0)) # number of velocities
@@ -465,7 +467,7 @@ class Roster():
                 
         return self._question_keys
     
-    def get_question_text(self, key):
+    def get_question_text(self, key) -> QuestionInfo:
         if key in self.question_keys().keys():
             return self.question_keys()[key]
         else:
@@ -658,7 +660,7 @@ class Roster():
         return self.stage_name_to_index_map.get(stage_name, 999)
     
     @property
-    def progress_summary(self):
+    def progress_summary(self) -> ProgressSummary:
         """Generated by Copilot"""
         if len(self.roster) == 0:
             return None
@@ -670,7 +672,7 @@ class Roster():
         }
     
     @property
-    def hubble_data(self):
+    def hubble_data(self) -> HubbleData:
         """Generated by Copilot"""
         summary = self.get_class_summary()
         return {
