+++ educator_dashboard/database/types.py
@@ -19,12 +19,6 @@ from .old_types import (
     OldStudentStoryState,
     StudentInfo,
     StudentEntry,
-    MeasurementStatus,
-    ShortReport,
-    QuestionInfo,
-    ClassSummary,
-    ProgressSummary,
-    HubbleData,
     ProcessedStage,
 )
